Make updateSection an action and remove boot helper
[MAILPOET-5088]
