--- mailpoet/assets/js/src/automation/integrations/mailpoet/analytics/helpers/boot.ts
@@ -1,12 +1,0 @@
-import { select } from '@wordpress/data';
-import { initializeApi, updateSection } from '../api';
-import { Section, storeName } from '../store';
-
-export function boot() {
-  initializeApi();
-  select(storeName)
-    .getSections()
-    .forEach((section: Section) => {
-      void updateSection(section);
-    });
-}
