+++ mailpoet/assets/js/src/automation/integrations/mailpoet/analytics/index.tsx
@@ -1,13 +1,14 @@
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
+import { dispatch, select } from '@wordpress/data';
 import { TopBarWithBeamer } from '../../../../common/top_bar/top_bar';
 import { Notices } from '../../../listing/components/notices';
 import { Header } from './components/header';
 import { Overview } from './components/overview';
 import { Tabs } from './components/tabs';
-import { createStore } from './store';
-import { boot } from './helpers/boot';
+import { createStore, Section, storeName } from './store';
 import { registerApiErrorHandler } from '../../../listing/api-error-handler';
+import { initializeApi } from './api';
 
 function Analytics(): JSX.Element {
   return (
@@ -27,6 +28,15 @@ function App(): JSX.Element {
       <Analytics />
     </BrowserRouter>
   );
+}
+
+function boot() {
+  initializeApi();
+  select(storeName)
+    .getSections()
+    .forEach((section: Section) => {
+      dispatch(storeName).updateSection(section);
+    });
 }
 
 window.addEventListener('DOMContentLoaded', () => {
